fix: miscellaneous bugs
