+++ app/build.gradle
@@ -48,10 +48,10 @@ dependencies {
 
 
     //otel
-    implementation("io.opentelemetry.instrumentation:opentelemetry-spring-boot-starter")
-    implementation("io.opentelemetry:opentelemetry-exporter-jaeger:1.34.1")
-    implementation 'io.opentelemetry.javaagent.instrumentation:opentelemetry-javaagent-redisson-3.17:2.9.0-alpha'
-
+    implementation("io.opentelemetry.instrumentation:opentelemetry-spring-boot-starter:2.6.0")
+    implementation 'ch.qos.logback:logback-classic:1.5.12'
+    implementation("io.opentelemetry:opentelemetry-exporter-jaeger:1.30.1")
+    implementation 'io.opentelemetry.javaagent.instrumentation:opentelemetry-javaagent-redisson-common:2.6.0-alpha'
     implementation 'io.netty:netty-resolver-dns-native-macos:4.1.68.Final:osx-aarch_64'
 
     // p6spy
@@ -79,6 +79,14 @@ dependencies {
 
 }
 
+//configurations {
+//    all {
+//        exclude group: 'org.springframework.boot', module: 'spring-boot-starter-logging'
+//        exclude group: 'ch.qos.logback', module: 'logback-classic'
+//        exclude group: 'org.apache.logging.log4j', module: 'log4j-to-slf4j'
+//    }
+//}
+
 //test {
 //    useJUnitPlatform()
 //}
@@ -88,21 +96,14 @@ tasks.named('test') {
 }
 
 tasks.withType(JavaCompile) {
-    options.compilerArgs += ["-parameters"]
     sourceCompatibility = '17'
     targetCompatibility = '17'
 }
 
-//configurations {
-//    all {
-//        exclude group: 'org.springframework.boot', module: 'spring-boot-starter-logging'
-//        exclude group: 'ch.qos.logback', module: 'logback-classic'
-//        exclude group: 'org.apache.logging.log4j', module: 'log4j-to-slf4j'
-//    }
-//}
 
 def swaggerRoot = "$projectDir/src/main/codegen-config".toString()
 def swaggerAPIOutput = "$projectDir/src/main".toString()
+
 
 
 openApiValidate {
