+++ app/src/main/java/com/joonhyeok/app/common/config/redis/RedisCacheConfig.java
@@ -1,12 +1,6 @@
 package com.joonhyeok.app.common.config.redis;
 
-import com.fasterxml.jackson.databind.DeserializationFeature;
-import com.fasterxml.jackson.databind.MapperFeature;
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.fasterxml.jackson.databind.SerializationFeature;
-import com.fasterxml.jackson.databind.json.JsonMapper;
-import com.fasterxml.jackson.databind.jsontype.BasicPolymorphicTypeValidator;
-import com.fasterxml.jackson.datatype.jsr310.JavaTimeModule;
 import com.joonhyeok.app.concert.application.dto.PerformanceDatesQueryResult;
 import com.joonhyeok.app.concert.application.dto.SeatsQueryResult;
 import org.springframework.cache.CacheManager;
@@ -25,31 +19,13 @@ import java.time.Duration;
 @Configuration
 @EnableCaching
 public class RedisCacheConfig {
-    @Bean
-    public ObjectMapper objectMapper() {
-        BasicPolymorphicTypeValidator validator = BasicPolymorphicTypeValidator.builder()
-                .allowIfSubType(Object.class)
-                .build();
-
-        ObjectMapper objectMapper = new JsonMapper();
-        objectMapper
-                .registerModule(new JavaTimeModule())
-                .setPolymorphicTypeValidator(validator)
-                .disable(SerializationFeature.WRITE_DATES_AS_TIMESTAMPS)
-                .configure(MapperFeature.USE_ANNOTATIONS, false)
-                .configure(MapperFeature.USE_GETTERS_AS_SETTERS, false)
-                .configure(DeserializationFeature.FAIL_ON_UNKNOWN_PROPERTIES, false);
-
-        return objectMapper;
-    }
-
 
     @Bean
     public CacheManager contentCacheManager(RedisConnectionFactory cf, ObjectMapper objectMapper) {
         RedisCacheConfiguration redisCacheConfiguration = RedisCacheConfiguration.defaultCacheConfig()
                 .serializeKeysWith(RedisSerializationContext.SerializationPair.fromSerializer(new StringRedisSerializer()))
-                .serializeValuesWith(RedisSerializationContext.SerializationPair.fromSerializer(new Jackson2JsonRedisSerializer<>(PerformanceDatesQueryResult.class)))
-                .serializeValuesWith(RedisSerializationContext.SerializationPair.fromSerializer(new Jackson2JsonRedisSerializer<>(SeatsQueryResult.class)))
+                .serializeValuesWith(RedisSerializationContext.SerializationPair.fromSerializer(new Jackson2JsonRedisSerializer<>(objectMapper, PerformanceDatesQueryResult.class)))
+                .serializeValuesWith(RedisSerializationContext.SerializationPair.fromSerializer(new Jackson2JsonRedisSerializer<>(objectMapper, SeatsQueryResult.class)))
                 .entryTtl(Duration.ofMinutes(30L));
 
         return RedisCacheManager.RedisCacheManagerBuilder.fromConnectionFactory(cf).cacheDefaults(redisCacheConfiguration).build();
