+++ app/src/main/java/com/joonhyeok/app/reservation/presentation/ReservationController.java
@@ -1,5 +1,6 @@
 package com.joonhyeok.app.reservation.presentation;
 
+import com.joonhyeok.app.common.aop.token.VerifyWait;
 import com.joonhyeok.app.reservation.application.MakeReservationService;
 import com.joonhyeok.app.reservation.application.dto.MakeReservationCommand;
 import com.joonhyeok.app.reservation.application.dto.MakeReservationResult;
@@ -10,6 +11,8 @@ import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestHeader;
 import org.springframework.web.bind.annotation.RestController;
 
 @Slf4j
@@ -18,9 +21,11 @@ import org.springframework.web.bind.annotation.RestController;
 public class ReservationController implements ReservationApi {
     private final MakeReservationService makeReservationService;
 
-    //    @VerifyWait
+    @VerifyWait
     @Override
-    public ResponseEntity<ReservationResponse> makeReservation(String waitToken, MakeReservationRequest request) {
+    public ResponseEntity<ReservationResponse> makeReservation(
+            @RequestHeader("Wait-Token") String waitToken,
+            @RequestBody MakeReservationRequest request) {
         MakeReservationResult result = makeReservationService.reserve(
                 new MakeReservationCommand(
                         request.getConcertId(),
